fix: add yargs, remove git fetch
